Removed sshd_config file
This file is an oversight and should not be here. The file is adding the
option "UseDNS no" to the base config and while this is an option that
could be set, this is not something that this role needs or should be
doing.  This issue was pointed out in the course of Bug: #1549379 being
worked on.

Change-Id: I4c184b0967d67e431b5b4e33172522c49459a3eb
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>