Tests: Make containers use the host's apt/pip config
This patch sets the test to check if the test is running in
OpenStack-CI, and if it is - then make the containers use
the host's pip.conf and some of the apt configuration.

All containers should make use of the hosts's apt sources
configuration. The tasks detect and implement this.

The intent is to make use of OpenStack-CI resources better
and consequently reduce the test run time.

Change-Id: I4a4ae2bdb5ffb909e087291d9198f0786e3f128a