Add insist option to lsyncd
Without insist, lsyncd will quit if any of its slaves are unavailable at
startup. This enforces a strict restart ordering - the master repo
server must come up last.

In order to remove this restriction the -insist option is added to
lsyncd. This causes lsyncd to retry the initial connection to each slave
so that even if the master comes up first, connections will eventually
be established.

Change-Id: I12e3dca147b5cb25ed982d5aceb9d521728c4e53
Closes-Bug: #1572433