Disable slave git servers while syncing
Now that the git repository is served from a different process
it is possible for there to be a race condition where the slaves
do not yet have the updated copy of the data before they get a
request for it.

This patch ensures that the git service is shut down on the slave
nodes during the synchronisation of repo data.

Change-Id: I83650253924354d0e63621474ba69f8cf254bde8