Cater for git cache when preparing repo server
When a git cache is implemented, a symlink is already in place
for the git folder when the repo server is prepared. This causes
the directory setup task to fail because the folder already
exists as a link.

This patch detects the presence of the folder and adjusts the
directory creation task to know that it should be a symlink.

The task is re-organised a bit to make it easier to read.

Change-Id: Idca73d6ac4dc3c668b6991ab2aae4784d1fe059d