Fix git service folder create task
The attempt to resolve the issue of a pre-existing
link in the place of the directory
( https://review.openstack.org/367563 ) did not work
as the file module requires a source and dest when
implementing a link.

This patch separates the task and only executes the
directory create if there is no link in place.

As per https://github.com/willthames/ansible-lint/issues/161
the mode for all directories is set to include a leading
zero to ensure that it's enforced as octal.

Change-Id: I27a64e8edaee1f652a4b3a95d05941df34824660