Ensure the repo server has a valid log directory
This change is simply making sure that the repo server has a valid
log directory at "/var/log/nginx". In some cases, generally when
upgrading, this directory may not exists and because the package
is already installed the directory will not be recreated. This
change simply makes sure the log directory is created regardless.

Closes-Bug: #1623778
Change-Id: I823aff32417804d494c46b0d444a518e2601118d
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>