Remove config tag for an install task
The rsyncd.defaults is not templated, and is the same for every
deploy. It should be allowed to be artifacted, and is not a config
task.

This cleans up the tag, and therefore will use the upper tag for
repo_pre_install.yml include, ``install``.

Change-Id: I1e98051cab01acc9b3dc505c1588c7916cbb2407
(cherry picked from commit 8a9c59585fabf94fbb0608912edb4055072ab382)