Remove sylinked git cache functionality
Prior to zuul v3 there was a cache of all git repositories
on the host, and this functionality used a bind-mount and
symlinks to pull the cache into the repo server as a way
to cut time taken and reduce errors due to git cloning.

The cache is no longer there, so the functionality which
implements the symlink and bind-mount is being removed.

The ability to make use of a git cache is still available,
but the cache will be synchronised into the container
instead of using bind mounts and symlinks.

This cuts down the complexity of the implementation.

Change-Id: I397c67a4b0166ad9f42a186f5c7137571772693a
Depends-On: Id95ad2f01b52222a07cb0775026d941ab86cd9ab