Fix swift port conflict
To avoid any port conflict, whether on LB or on bare metal nodes,
we should make sure each service runs on its own port.

The 8080 port is already used by swift, and opendaylight.
We keep 8080 for swift, move opendaylight to 8180, and the
pypiserver to 8280 to avoid overlaps when everything runs on metal.

Closes-Bug: 1735764
Change-Id: I69dd043efe5d2e50e83014bdbd6a848bfcc2aa39