Revert "Remove sylinked git cache functionality"
This reverts commit 9c1859c7f98c16a6f5d480b542a986df01fd499b.
For deployers having symlinks in the P branch, the removal of
those tasks would be problematic when doing an upgrade to Q.

Instead, we keep those tasks, and mention their deprecation.
People wanting to use the git cache can still use it, but
as a plain folder instead of a symlink.

Change-Id: Ief7ad423b9e701c145b97de42e5165fe2f7347d7