If virtualenv is broken, force re-install it
The virtualenv_support folder missing from the repo
container when upgrading due to the implementation
in the repo_build role which removes it. In order
to prevent breaking the virtualenv tooling, [1] is
being submitted. However, we still need to handle
the situation where the upgrade is done from a
version before that patch is implemented.

This patch implements a force-reinstallation of
virtualenv if the folder is missing.

[1] https://review.openstack.org/540351

Change-Id: I2dcfa1abb46ee4e088dac7c9e77e9772ba8a269e
Closes-Bug: #1746935