Allow pypiserver to be upgraded before repo build
To cater for a situation where the pypiserver package is changing,
but the repo does not yet have the package built, we need to ensure
that the install task for it can fetch from pypi (or the designated
pypi mirror). To do this we try the local repo first, then fall back
to using the designated default index which defaults to pypi.

Change-Id: I1a2d36793fccc4c50f00247c8d19501ad8816517
(cherry picked from commit 4e3213f4c6618235aa93bd994166799c06ccb7a5)