Remove the upstream pypi reverse proxy
Trying to reverse proxy upstream pypi has not turned out to
be very stable, or very useful. We've had many, many reports
of stability issues and the additional complexity for offline
and proxy usage is just not worth it.

Given we already have a mechanism in place to handle using
upstream pypi if the repo server is not there yet, disabling
this should just result in that mechanism kicking in and all
will be well again.

Once the repo is built, the reverse proxy to pypiserver will
then be exclusively used and the upstream pypi proxy is not
necessary anyway.

Depends-On: https://review.openstack.org/584394
Change-Id: Ie407c6a346de6b46c8f4d30caea8664a7f6bd341
(cherry picked from commit dab934bdb9f06a4c194469afa34a0998b76721b6)