Use distro packages only after they are installed
In an environment with sudo and lsyncd not already installed the pre
install tasks try to set up config files before the config directories
exist, and the tasks fail.

This patch moves the sudo and lsyncd config tasks to run after the
distro packages have been installed.

In addition, sudo is added as a required distro package for the
repo server.

Change-Id: Iab09b732d63d007c1d638ec9dc35834e564e5ce5