Major Upgrade: Remove pypiserver from repo server
In https://review.opendev.org/648556 we release noted the removal of
the pypiserver service, but nothing has actually removed it during
the upgrade process.

This patch implements the removal. This is implemented as a handler
to ensure that the service and nginx config are removed only when
the changes are made to the service restarts.

Change-Id: I06367f3448a4d2bd932d868cd7ea825287b84fbc