Have a symlink to u_c versioned file
With current behavior of u-c file that is stored on repo server and
migration to shared FS instead of lsyncd, we don't have any reliable
path to check for.

Also we have issue, that with updated u-c SHA users catch "unrelated"
error in python_venv_build role.

As a solution, we can create a symlink to the u-c file, that will have
persistant name and be updated with repo_server role run.

That would give us both file to verify and it can be used as u-c URI.

Change-Id: Ie7bdb9137ed69be465f014fa811b750dbec4e428
Related-Bug: #1943978