Restart nginx after removing old repo content
The version of nginx on centos-8 appears to keep file handles open,
possibly the old /var/www/repo directory persistently. Once the old
content is removed and the new shared filesystem mount is created at
the web root, ensure that nginx is restarted to close any file handles
which are now stale.

Change-Id: I941359b1b42aa4a874230a32b438dcefddfb2acb