Ensure mounts are present only when they are expected to exist
At the moment there is a check, which ensures for race condition to
not happen between asking for a mount and mounting the point.
However, it tries to check for the mount when there is no mount
defined for the directory.

We add extra check to wait for the mount only when mounts are defined.

Change-Id: I900a55a6f4edce3d3fe419821c47cf56d641192f