Increase timeouts and add retries for UC fetching over HTTP
In CI we've spotted multiple occurances of connection timeouts towards
Gitea resulting in intermittent failures.

Default timeout for get_url module is 10 seconds, which might be not
enough for fetching the file.
A retry is also being added here as another layer of protection for
remote operations
to reduce risk of intermittent failures.

Change-Id: I42663e8316ef87c12a229517ac0f95f8c6c2c14b