Remove access limitations to repo vhost
With migration from Nginx to Apache [1] an IP limitation was introduced
which was not present before.
It is based on asumption that container_networks are defined which is
not the case for some metal deployments.

We need to make a fix backportable and focus of the change was migraton
from Nginx to Apache rather then change behaviour.

With that in mind revert of the behaviour change is proposed, while
re-implementation
of limitation can be re-implemented later with a follow-up patch.

[1] https://review.opendev.org/c/openstack/openstack-ansible-repo_server/+/924162

Closes-Bug: #2093011
Change-Id: Ic4f4405ad7ffe1427791a012bb56f8fefa16d145